validate ELF LSB executable

+++ ukonvrt/check-app-osv-fit.sh
@@ -4,18 +4,29 @@
 
 OSV_BASE=/osv-base/loader.elf
 
-# Adapted from
-# https://github.com/cloudius-systems/osv/blob/master/scripts/check-libcfunc-avail.sh
-echo "Checking if application GLIBC symbols are available from OSv.."
-DUMPFILE=`mktemp`
-objdump -t $OSV_BASE > $DUMPFILE
-FUNCS=`objdump -T $APP | grep GLIBC|sed -e "s/.*GLIBC\(XX\)\?_[0-9.]* //"`
+FILE_TYPE=`file $APP | awk -F[:,] '{print $2}'`
+echo "$APP is of type $FILE_TYPE"
 
-for i in $FUNCS; do
-    grep -q " $i$" $DUMPFILE
-    FOUND=$?
-    if [ $FOUND != 0 ]; then
-        echo "$i not found"
-    fi
-done
-rm $DUMPFILE
+if [ "$FILE_TYPE" == " ELF 64-bit LSB executable" ]; then
+    # 64-bit System V ELF payload
+
+    # Adapted from
+    # https://github.com/cloudius-systems/osv/blob/master/scripts/check-libcfunc-avail.sh
+    echo "Checking if application GLIBC symbols are available from OSv.."
+    DUMPFILE=`mktemp`
+    objdump -t $OSV_BASE > $DUMPFILE
+    FUNCS=`objdump -T $APP | grep GLIBC|sed -e "s/.*GLIBC\(XX\)\?_[0-9.]* //"`
+
+    for i in $FUNCS; do
+        grep -q " $i$" $DUMPFILE
+        FOUND=$?
+        if [ $FOUND != 0 ]; then
+            echo "$i not found"
+        fi
+    done
+    rm $DUMPFILE
+
+else
+    echo "$FILE_TYPE is currently unsupported"
+fi
+
